Conectado componente ListaPinosQueHablaron a los oradores de un tema
